Move non-parametric controlled ops inherit from Controlled
