+++ pennylane/ops/op_math/__init__.py
@@ -64,6 +64,9 @@ Controlled Operator Classes
     ~ControlledQubitUnitary
     ~CY
     ~CZ
+    ~CH
+    ~CNOT
+    ~Toffoli
     ~CRX
     ~CRY
     ~CRZ
@@ -109,6 +112,11 @@ from .controlled_ops import (
     CRZ,
     CY,
     CZ,
+    CH,
+    CCZ,
+    CSWAP,
+    CNOT,
+    Toffoli,
     MultiControlledX,
 )
 from .decompositions import one_qubit_decomposition, two_qubit_decomposition, sk_decomposition
@@ -125,6 +133,11 @@ controlled_qubit_ops = {
     "ControlledQubitUnitary",
     "CY",
     "CZ",
+    "CH",
+    "CCZ",
+    "CSWAP",
+    "CNOT",
+    "Toffoli",
     "MultiControlledX",
     "CRX",
     "CRY",
