+++ pennylane/ops/qubit/__init__.py
@@ -57,17 +57,12 @@ __ops__ = {
     "S",
     "T",
     "SX",
-    "CNOT",
-    "CCZ",
-    "CH",
     "SWAP",
     "ISWAP",
     "SISWAP",
     "SQISW",
-    "CSWAP",
     "PSWAP",
     "ECR",
-    "Toffoli",
     "RX",
     "RY",
     "RZ",
