+++ tests/drawer/test_drawer_utils.py
@@ -137,8 +137,8 @@ class TestUnwrapControls:
         "op,expected_control_wires,expected_control_values",
         [
             (qml.PauliX(wires="a"), Wires([]), None),
-            (qml.CNOT(wires=["a", "b"]), Wires("a"), None),
-            (qml.ctrl(qml.PauliX(wires="b"), control="a"), Wires("a"), None),
+            (qml.CNOT(wires=["a", "b"]), Wires("a"), [True]),
+            (qml.ctrl(qml.PauliX(wires="b"), control="a"), Wires("a"), [True]),
             (
                 qml.ctrl(qml.PauliX(wires="b"), control=["a", "c", "d"]),
                 Wires(["a", "c", "d"]),
