+++ tests/ops/op_math/test_pow_op.py
@@ -464,12 +464,11 @@ class TestSimplify:
 
     def test_simplify_method_with_controlled_operation(self):
         """Test simplify method with controlled operation."""
-        pow_op = Pow(ControlledOp(base=qml.Hadamard(0), control_wires=1, id=3), z=3)
-        final_op = qml.CH([1, 0], id=3)
+        pow_op = Pow(ControlledOp(base=qml.Hadamard(0), control_wires=1), z=3)
+        final_op = qml.CH([1, 0])
         simplified_op = pow_op.simplify()
 
-        # TODO: uncomment this check when all controlled operations inherit from ControlledOp
-        # assert isinstance(simplified_op, ControlledOp)
+        assert isinstance(simplified_op, ControlledOp)
         assert final_op.data == simplified_op.data
         assert final_op.wires == simplified_op.wires
         assert final_op.arithmetic_depth == simplified_op.arithmetic_depth
